+++ src/helpers/ObjectHelpers.spec.js
@@ -1,7 +1,10 @@
 import * as objectHelpers from "./ObjectHelpers";
 
 import irises from "../data/mocks/c88b3df0-de91-4f5b-a9ef-7b2b9a6d8abb.json";
+import beetle from "../data/mocks/a69d5696-70c2-56ed-9f82-fb2e69311e5d.json";
 import fables from "../data/mocks/f8fd6961-6da3-4c39-94ad-e8e9367fa51b.json";
+import kasten from "../data/mocks/1cd16038-091e-4010-ae07-df95acfce9a4.json";
+import triton from "../data/mocks/dff75e58-f8b9-4507-8ab7-5d948451dea7";
 
 describe("getDimensionsDescription", () => {
   it("gets the correct dimensions description if present", () => {
@@ -11,9 +14,49 @@ describe("getDimensionsDescription", () => {
   });
 });
 
-describe("getAccessionNumber", () => {
+describe("getAccessionNumbers", () => {
   it("gets the correct accession number if present", () => {
     expect(objectHelpers.getAccessionNumbers(irises)).toEqual(["90.PA.20"]);
+  });
+});
+
+describe("getDigitalImages", () => {
+  it("gets the correct image urls if present", () => {
+    expect(objectHelpers.getDigitalImages(irises)[0]).toEqual(
+      "https://media.getty.edu/iiif/image/e5d29650-11f8-4897-9540-54a9dd65b04f/full/full/0/default.jpg"
+    );
+  });
+});
+
+describe("getRightsStatements", () => {
+  it("gets the correct right statement if present", () => {
+    expect(objectHelpers.getRightsStatements(irises)).toEqual([
+      "No Copyright - United States",
+    ]);
+  });
+});
+
+describe("getCopyrightStatements", () => {
+  it("return empty array if no copyright", () => {
+    expect(objectHelpers.getCopyrightStatements(kasten)).toEqual([
+      "© Barbara Kasten",
+    ]);
+  });
+});
+
+describe("getRightsAssertions", () => {
+  it("gets the correct rightsAssertions if present", () => {
+    expect(objectHelpers.getRightsAssertions(beetle)[0]).toEqual(
+      "https://rightsstatements.org/vocab/NoC-US/1.0/"
+    );
+  });
+});
+
+describe("getAcknowledgements", () => {
+  it("gets the correct acknowledgements if present", () => {
+    expect(objectHelpers.getAcknowledgementStatements(triton)).toEqual(
+      undefined
+    );
   });
 });
 
@@ -59,6 +102,12 @@ describe("get creators", () => {
         label: "Berengarius Fredoli",
       },
     ]);
+  });
+});
+
+describe("getCultures", () => {
+  it("gets the correct culture(s) if present", () => {
+    expect(objectHelpers.getCultures(irises)).toEqual(["Dutch"]);
   });
 });